+++ R/globals.R
@@ -1,1 +1,1 @@
-utils::globalVariables(c("value", "variable", "time", ".", "x", "S", "E", "I", "R"))
+utils::globalVariables(c("value", "variable", "time", ".", "x", "S", "E", "I", "R", "hcl"))